+++ src/main/webapp/components/gallery/gallery.js
@@ -4,7 +4,7 @@ import Dialog from '@material-ui/core/Dialog'
 import Typography from '@material-ui/core/Typography'
 import React from 'react'
 import Inspector from '../inspector'
-import Thumbnail from '../../thumbnail/thumbnail'
+import Thumbnail from '../thumbnail/thumbnail'
 
 const GalleryItem = props => {
   const { title, thumbnail } = props
