+++ src/main/webapp/components/thumbnail/index.js
@@ -1,0 +1,3 @@
+import Thumbnail from './thumbnail'
+
+export { Thumbnail as default }
