+++ src/main/webapp/components/thumbnail/thumbnail.js
