+++ src/main/webapp/search/details.js
@@ -8,7 +8,7 @@ import gql from 'graphql-tag'
 import { getIn } from 'immutable'
 import React from 'react'
 import { useParams } from 'react-router-dom'
-import Thumbnail from '../thumbnail/thumbnail'
+import Thumbnail from '../components/thumbnail/thumbnail'
 import ErrorMessage from '../components/network-retry/inline-retry'
 
 const LoadingComponent = () => <LinearProgress />
