+++ src/main/webapp/search/results.js
@@ -14,7 +14,7 @@ import TextFieldsIcon from '@material-ui/icons/TextFields'
 import gql from 'graphql-tag'
 import React from 'react'
 import ErrorMessage from '../components/network-retry/inline-retry'
-import Thumbnail from '../thumbnail/thumbnail'
+import Thumbnail from '../components/thumbnail/thumbnail'
 import Search from './search'
 import useSearchRouting, { DetailsLink } from './search-routing'
